Fix errors when building for a 32-bit x86 target.
Use size_t directly where possible, or cast to ptrdiff_t when required
as an index into an iterator.

Decorate methods bound to JavaScript with JSC_HOST_CALL to match the
required "fastcall" method signature.

+++ cobalt/speech/microphone_fake.cc
@@ -100,7 +100,8 @@ bool MicrophoneFake::Open() {
   if (read_data_from_file_) {
     // Read from local files.
     SB_DCHECK(file_paths_.size() != 0);
-    uint64 random_index = base::RandGenerator(file_paths_.size());
+    size_t random_index =
+        static_cast<size_t>(base::RandGenerator(file_paths_.size()));
     starboard::ScopedFile file(file_paths_[random_index].value().c_str(),
                                kSbFileOpenOnly | kSbFileRead, NULL, NULL);
     SB_DCHECK(file.IsValid());
